requirements.txt
ass==0.4.4
astroid==2.2.5
autopep8==1.4.3
certifi==2019.3.9
chardet==3.0.4
Django==2.1.7
django-cleanup==3.2.0
django-haystack==2.8.1
django-sslserver==0.20
django-taggit==1.1.0
isort==4.3.16
jieba==0.39
lazy-object-proxy==1.3.1
mccabe==0.6.1
pycodestyle==2.5.0
pylint==2.3.1
PySocks==1.6.8
python-magic-bin==0.4.14
pytz==2018.9
six==1.12.0
sqlparse==0.3.0
typed-ast==1.3.1
Whoosh==2.7.4
wrapt==1.11.1
